fix(): GastoRepository, GastoService, Mascota.
Ahora se borra correctamente

+++ src/main/java/mivet/service/GastoService.java
@@ -1,14 +1,17 @@
 package mivet.service;
 
 import mivet.model.Gasto;
+import mivet.model.Mascota;
 import mivet.repository.GastoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@Transactional
 public class GastoService {
 
     private final GastoRepository gastoRepository;
@@ -31,8 +34,19 @@ public class GastoService {
     }
 
     public void delete(Long id) {
+    Optional<Gasto> gastoOpt = gastoRepository.findById(id);
+    if (gastoOpt.isPresent()) {
+        Gasto gasto = gastoOpt.get();
+        Mascota mascota = gasto.getMascota();
+
+        if (mascota != null) {
+            mascota.getGastos().remove(gasto);
+        }
+
         gastoRepository.deleteById(id);
     }
+}
+
 
     public List<Gasto> findByMascotaId(Long mascotaId) {
         return gastoRepository.findByMascotaId(mascotaId);
